Introduce Artist.findLatestByMbid
This introduces the CoreEntity type, which contains a version a MusicBrainz
core entity. That is, something which is versioned, has data, and has an
MBID. The reference to a core entity is a Revision, and core entitys also
provide an MBID to data which is related but not versioned by this entity
(for example, release groups for an artist).

findLatestByMbid finds an artist by finding the latest revision for a given
MBID.